+++ response.h
@@ -1,7 +1,9 @@
+#include <stdio.h>
+
 typedef struct Response {
 	int status;
 	char header[1024];
 	char body[6144];
 } Response;
 
-int response(char (*buffer)[8192], Response *res);
+size_t response(char (*buffer)[8192], Response *res);
